+++ src/SubscriptionItem.php
@@ -40,7 +40,7 @@ class SubscriptionItem extends Model
      */
     public function subscription()
     {
-        return $this->belongsTo(Subscription::class);
+        return $this->belongsTo(Cashier::$subscriptionModel);
     }
 
     /**